Refactor Analyzer by moving checks into /plugins folder
